move groups-related components to dedicated Groups component folder
